+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Fix isomorphic-fetch import
+- Fix getSpecificBuild not always returning a Promise
 
 ## [0.3.0] - 2017-03-23 [YANKED]
 ### Added