+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix isomorphic-fetch import
 - Fix getSpecificBuild not always returning a Promise
+- Fix missing accept header
 
 ## [0.3.0] - 2017-03-23 [YANKED]
 ### Added